chore: organize ci-v5 unit tests
