+++ packages/ci-v5/package copy.json
@@ -1,0 +1,55 @@
+{
+  "name": "@heroku-cli/plugin-certs",
+  "version": "8.1.4",
+  "author": "Jeff Dickey @jdxcode",
+  "bugs": "https://github.com/heroku/cli/issues",
+  "dependencies": {
+    "@oclif/core": "^1.26.2",
+    "tslib": "^1.9.3"
+  },
+  "devDependencies": {
+    "@oclif/test": "^2.2.20",
+    "@types/chai": "^4.1.7",
+    "@types/mocha": "^5.2.6",
+    "chai": "^4.2.0",
+    "globby": "^10.0.2",
+    "mocha": "^5",
+    "nyc": "^15.1.0",
+    "oclif": "3.8.1",
+    "ts-node": "^10.9.1",
+    "typescript": "4.8.4"
+  },
+  "engines": {
+    "node": ">=14"
+  },
+  "files": [
+    "/lib",
+    "/oclif.manifest.json"
+  ],
+  "homepage": "https://github.com/heroku/cli/tree/main/packages/certs",
+  "keywords": [
+    "oclif-plugin"
+  ],
+  "license": "MIT",
+  "oclif": {
+    "additionalHelpFlags": [
+      "-h"
+    ],
+    "additionalVersionFlags": [
+      "-v"
+    ],
+    "commands": "./lib/commands",
+    "bin": "oclif-example",
+    "repositoryPrefix": "<%- repo %>/blob/v<%- version %>/packages/certs/<%- commandPath %>"
+  },
+  "repository": "heroku/cli",
+  "scripts": {
+    "lint": "eslint . --ext .ts --config ../../.eslintrc --ignore-path ../../.eslintignore-lib",
+    "postpack": "rm -f oclif.manifest.json",
+    "prepack": "rm -rf lib && tsc && oclif manifest",
+    "pretest": "tsc -p test --noEmit",
+    "test": "nyc mocha --forbid-only \"test/**/*.unit.test.js\"",
+    "posttest": "yarn lint",
+    "version": "oclif readme && git add README.md"
+  }
+}
