+++ packages/ci-v5/test/unit/commands/ci/config-set.unit.test.js
@@ -3,7 +3,7 @@
 const nock = require('nock')
 const expect = require('chai').expect
 const cli = require('heroku-cli-util')
-const cmd = require('../../../commands/ci/config-set')
+const cmd = require('../../../../commands/ci/config-set')
 const Factory = require('../../lib/factory')
 
 describe('heroku ci:config:set', function () {
