+++ packages/ci-v5/test/unit/lib/heroku-api.unit.test.js
@@ -4,7 +4,7 @@
 const nock = require('nock')
 const expect = require('chai').expect
 const Heroku = require('heroku-client')
-const herokuAPI = require('../../lib/heroku-api')
+const herokuAPI = require('../../../lib/heroku-api')
 
 describe('heroku-api', function () {
   afterEach(() => nock.cleanAll())
